Move functions to the package-level
