+++ src/jsMain/kotlin/KtUniversalis.js.kt
@@ -1,0 +1,17 @@
+package cloud.drakon.ktuniversalis
+
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.js.Js
+import io.ktor.client.plugins.DefaultRequest
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.serialization.kotlinx.json.json
+
+internal actual val ktorClient = HttpClient(Js) {
+    install(ContentNegotiation) {
+        json()
+    }
+
+    install(DefaultRequest) {
+        url("https://universalis.app/api/v2/")
+    }
+}
