+++ src/jvmMain/kotlin/KtUniversalis.jvm.kt
@@ -1,0 +1,17 @@
+package cloud.drakon.ktuniversalis
+
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.java.Java
+import io.ktor.client.plugins.DefaultRequest
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.serialization.kotlinx.json.json
+
+internal actual val ktorClient = HttpClient(Java) {
+    install(ContentNegotiation) {
+        json()
+    }
+
+    install(DefaultRequest) {
+        url("https://universalis.app/api/v2/")
+    }
+}
